feat(streams): stabilize
