+++ encoding/unstable_base64_stream.ts
@@ -138,7 +138,7 @@ export class Base64EncoderStream<T extends "string" | "bytes">
  *   Base64DecoderStream,
  *   Base64EncoderStream,
  * } from "@std/encoding/unstable-base64-stream";
- * import { toBytes } from "@std/streams/unstable-to-bytes";
+ * import { toBytes } from "@std/streams/to-bytes";
  *
  * const readable = (await Deno.open("./deno.lock"))
  *   .readable
