+++ streams/deno.json
@@ -19,7 +19,7 @@
     "./to-array-buffer": "./to_array_buffer.ts",
     "./to-blob": "./to_blob.ts",
     "./unstable-to-byte-stream": "./unstable_to_byte_stream.ts",
-    "./unstable-to-bytes": "./unstable_to_bytes.ts",
+    "./to-bytes": "./to_bytes.ts",
     "./to-json": "./to_json.ts",
     "./unstable-to-lines": "./unstable_to_lines.ts",
     "./to-text": "./to_text.ts",
